# Speed up CUDA graph capture on Flashinfer