--- a/python/sglang/srt/model_executor/cuda_graph_runner.py
+++ b/python/sglang/srt/model_executor/cuda_graph_runner.py
@@ -624,6 +624,16 @@ class CudaGraphRunner:
             self.model_runner.lora_manager.prepare_lora_batch(forward_batch)
 
         # Attention backend
+        # Provide CPU seq_lens to attn backend(s) for capture to avoid extra .cpu() copies
+        try:
+            setattr(self.model_runner.attn_backend, "_capture_seq_lens_cpu", self.seq_lens_cpu)
+            # In speculative/multistep backends, propagate to nested backends if any
+            if hasattr(self.model_runner.attn_backend, "attn_backends"):
+                for ab in getattr(self.model_runner.attn_backend, "attn_backends"):
+                    setattr(ab, "_capture_seq_lens_cpu", self.seq_lens_cpu)
+        except Exception:
+            pass
+
         self.model_runner.attn_backend.init_forward_metadata_capture_cuda_graph(
             bs,
             num_tokens,