<!-- Thank you for your contribution! Please follow these guidelines to enhance your pull request. If anything is unclear, submit your PR and reach out to maintainers for assistance. Join our Slack community at https://slack.sglang.ai to discuss further. -->

## Motivation

After #8340, try to speed up the graph capture by removing the host->device sync.

@DarkSharpness 

<!-- Describe the purpose and goals of this pull request. -->

## Benchmarking and Profiling

On A100-SXM4-40GB

## Before

```bash
SGLANG_ENABLE_TORCH_COMPILE=0 python3 -m sglang.launch_server \
    --attention-backend flashinfer \
    --model-path Qwen/Qwen2.5-7B-Instruct

Capture cuda graph bs [1, 2, 4, 8, 16, 24, 32, 40, 48, 56, 64, 72, 80, 88, 96, 104, 112, 120, 128, 136, 144, 152, 160]
```

Cold start
```
[2025-08-25 15:31:27] Capture cuda graph end. Time elapsed: 25.39 s. mem usage=0.56 GB. avail mem=8.00 GB.
```

Non-cold start
```
[2025-08-25 15:32:45] Capture cuda graph end. Time elapsed: 3.57 s. mem usage=0.56 GB. avail mem=8.00 GB.
...
[2025-08-25 15:34:38] Capture cuda graph end. Time elapsed: 3.70 s. mem usage=0.56 GB. avail mem=8.00 GB.
```

---

## After

Cold start
```
[2025-08-25 15:40:06] Capture cuda graph end. Time elapsed: 24.57 s. mem usage=0.56 GB. avail mem=8.00 GB.
```

Non-cold start
```
[2025-08-25 15:41:27] Capture cuda graph end. Time elapsed: 3.40 s. mem usage=0.56 GB. avail mem=8.00 GB.
...
[2025-08-25 15:42:25] Capture cuda graph end. Time elapsed: 3.37 s. mem usage=0.56 GB. avail mem=8.00 GB.
```

## Checklist
- [x] Provide accuracy and speed benchmark results according to [Test the accuracy](https://docs.sglang.ai/developer_guide/contribution_guide.html#test-the-accuracy) and [Benchmark the speed](https://docs.sglang.ai/developer_guide/contribution_guide.html#benchmark-the-speed).